Create common.php
